+++ src/cli/commands/patch/apply.rs
@@ -14,8 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://gnu.org/licenses/gpl-3.0.html>.
 
-use bitcoin_hashes::Sha1;
 use clap::Args;
+use nostr::hashes::sha1::Hash as Sha1Hash;
 
 use super::PatchStatus;
 use crate::{
@@ -40,7 +40,7 @@ pub struct ApplyArgs {
     patch_id:        NostrEvent,
     /// The applied commits
     #[arg(num_args = 1..)]
-    applied_commits: Vec<Sha1>,
+    applied_commits: Vec<Sha1Hash>,
 }
 
 impl CommandRunner for ApplyArgs {